Add exception conditional|refactor print_board
To make it a little more readable and easy to understand from all
perspectives, I added a conditional when getting user input, so they
are only able to use the numbers 1-9. I needed a way to also have the
board only print a number on the board if there wasn't a token in the
cell already, so I used slice and map to print out the board with
separators.