Add check in tie? method
Since tie previously had a bug in it that would result in a tie if the
board was filled, even with a winning line, I originally just changed
the code to do a check for a win first. With this change, the check
also happens in the tie? method, for a more flexible usage.