Fix bad tests and create test for meth
I realized my test for testing win? and tie? were bad, as the way I was
calling them would test the wrong things. Now I am testing the methods
when they are called, and checking the specific value, which works as I
intended. The added method test was for #spot_empty?, which I utilized
the same style to get the correct result.